agents/linkedin_lookup_agent.py
import os
from dotenv import load_dotenv

load_dotenv()

from langchain_openai import ChatOpenAI
from langchain_core.prompts import PromptTemplate
from langchain_core.tools import Tool
from langchain.agents import (
    create_react_agent,
    AgentExecutor,
)

def lookup(name: str) -> str:
    return "https://www.linkedin.com/in/eden-marco/"