requirements.txt
langchain-core~=0.3.49
langchain-openai~=0.3.11
h11~=0.14.0
pip~=24.3.1
distro~=1.9.0
filelock~=3.18.0
requests~=2.32.3
PyYAML~=6.0.2
sniffio~=1.3.1
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
jiter~=0.9.0
regex~=2024.11.6
typing_extensions~=4.13.0
openai~=1.69.0
pydantic~=2.11.1
anyio~=4.9.0
pydantic_core~=2.33.0
orjson~=3.10.16
certifi~=2025.1.31
urllib3~=2.3.0
annotated-types~=0.7.0
charset-normalizer~=3.4.1
tenacity~=9.0.0
tiktoken~=0.9.0
langsmith~=0.3.19
requests-toolbelt~=1.0.0
packaging~=24.2
jsonpointer~=3.0.0
jsonpatch~=1.33
python-dotenv~=1.1.0
platformdirs~=4.3.7
ollama~=0.4.7
langchain~=0.3.21
langchain-text-splitters~=0.3.7
SQLAlchemy~=2.0.40